% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylima-package.R
\docType{package}
\name{shinylima-package}
\alias{shinylima}
\alias{shinylima-package}
\title{shinylima: LIMA-Tool}
\description{
Die Revision des Grundangebot zum Thema Hand채nderungen f체hrt auch zu Ver채nderungen in der LIMA-App: Neu werden Bodenpreise und Wohnfl채chenpreise in unterschiedlichen Apps ausgewiesen. Gleichzeitig wird der Aufbau der App modularisiert.
}
\author{
\strong{Maintainer}: Aline Metzler \email{aline.metzler@zuerich.ch}

}
\keyword{internal}
